update inventory controllers and migrate frontend inventory API calls to axiosInstance
