+++ petclinic-frontend/src/features/inventories/InventoriesListTable.tsx
@@ -11,6 +11,8 @@ import AddInventoryType from '@/features/inventories/AddInventoryType.tsx';
 import { ProductModel } from '@/features/inventories/models/ProductModels/ProductModel.ts';
 import inventoryStyles from './InventoriesListTable.module.css';
 import cardStylesInventory from './CardInventoryTeam.module.css';
+// import axios from 'axios';
+import axiosInstance from '@/shared/api/axiosInstance';
 
 export default function InventoriesListTable(): JSX.Element {
   const [selectedInventories, setSelectedInventories] = useState<Inventory[]>(
@@ -116,24 +118,16 @@ export default function InventoriesListTable(): JSX.Element {
 
   const fetchProductQuantity = async (inventoryId: string): Promise<void> => {
     try {
-      const response = await fetch(
-        `http://localhost:8080/api/v2/gateway/inventories/${inventoryId}/productquantity`,
-        {
-          method: 'GET',
-          credentials: 'include', // <-- Add this line to include cookies or credentials
-          headers: { 'Content-Type': 'application/json' },
-        }
+      const response = await axiosInstance.get<number>(
+        `/inventory/${inventoryId}/productquantity`,
+        { useV2: false }
       );
 
-      if (response.ok) {
-        const quantity = await response.json();
-        setProductQuantities(prevQuantities => ({
-          ...prevQuantities,
-          [inventoryId]: quantity,
-        }));
-      } else {
-        console.error('Failed to fetch product quantity:', response.statusText);
-      }
+      const quantity = await response.data;
+      setProductQuantities(prevQuantities => ({
+        ...prevQuantities,
+        [inventoryId]: quantity,
+      }));
     } catch (error) {
       console.error('Error fetching product quantity:', error);
     }
@@ -143,30 +137,17 @@ export default function InventoriesListTable(): JSX.Element {
     inventory: Inventory
   ): Promise<void> => {
     try {
-      const response = await fetch(
-        `http://localhost:8080/api/gateway/inventory/${inventory.inventoryId}/products/lowstock`,
-        {
-          method: 'GET',
-          credentials: 'include',
-          headers: {
-            Accept: 'application/json',
-          },
-        }
+      const response = await axiosInstance.get<ProductModel[]>(
+        `/inventory/${inventory.inventoryId}/products/lowstock`,
+        { useV2: false }
       );
-
-      if (response.ok) {
-        const data = await response.json();
-        if (data && data.length > 0) {
-          // Update the ref directly
-          lowStockProductsByInventory.current = {
-            ...lowStockProductsByInventory.current,
-            [inventory.inventoryName]: data,
-          };
-        }
-      } else {
-        console.error(
-          `No products below threshold in inventory: ${inventory.inventoryName}`
-        );
+      const data = response.data;
+      if (data && data.length > 0) {
+        // Update the ref directly
+        lowStockProductsByInventory.current = {
+          ...lowStockProductsByInventory.current,
+          [inventory.inventoryName]: data,
+        };
       }
     } catch (error) {
       console.error('Error fetching low stock products:', error);
