+++ petclinic-frontend/src/features/inventories/api/addProductToInventoryByName.ts
@@ -6,9 +6,16 @@ export async function addProductToInventoryByName(
   inventoryName: string,
   product: ProductModelINVT
 ): Promise<InventoryName[]> {
-  const response = await axiosInstance.post(
-    `http://localhost:8080/api/v2/gateway/inventories/${inventoryName}/products/by-name`,
-    product
-  );
-  return response.data;
+  //Not implemented neither in v1 or v2
+  try {
+    const response = await axiosInstance.post(
+      `/inventory/${inventoryName}/products/by-name`,
+      product,
+      { useV2: false }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error adding Product to Inventory:', error);
+    throw error;
+  }
 }
