+++ petclinic-frontend/src/features/inventories/api/getAllInventoryNames.ts
@@ -2,8 +2,14 @@ import axiosInstance from '@/shared/api/axiosInstance.ts';
 import { InventoryName } from '@/features/inventories/models/InventoryName.ts';
 
 export async function getAllInventoryNames(): Promise<InventoryName[]> {
-  const response = await axiosInstance.get<InventoryName[]>(
-    axiosInstance.defaults.baseURL + 'inventories/names'
-  );
-  return response.data;
+  try {
+    const response = await axiosInstance.get<InventoryName[]>(
+      '/inventory/names',
+      { useV2: false } // not implemented
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error get all inventory by names:', error);
+    throw error;
+  }
 }
